requirements.txt
langchain-openai  
langchain-community  
langchain-core  
python-dotenv  
langgraph  
tavily-python  
streamlit  
pandas